Allow to disable ldap groups sync with env flag
