+++ modules/ldap_groups/app/workers/ldap_groups/synchronization_job.rb
@@ -35,32 +35,15 @@ module LdapGroups
 
     def perform
       return unless EnterpriseToken.allows_to?(:ldap_groups)
+      return if skipped?
 
       User.system.run_given do
-        synchronize!
+        ::LdapGroups::SynchronizationService.synchronize!
       end
     end
 
-    private
-
-    def synchronize!
-      LdapAuthSource.find_each do |ldap|
-        Rails.logger.info { "[LDAP groups] Retrieving groups from filters for ldap auth source #{ldap.name}" }
-        LdapGroups::SynchronizedFilter
-          .where(auth_source_id: ldap.id)
-          .find_each do |filter|
-
-          LdapGroups::SynchronizeFilterService
-            .new(filter)
-            .call
-        end
-
-        Rails.logger.info { "[LDAP groups] Start group synchronization for ldap auth source #{ldap.name}" }
-        LdapGroups::SynchronizationService.new(ldap).call
-      end
-    rescue StandardError => e
-      msg = "[LDAP groups] Failed to run LDAP group synchronization. #{e.class.name}: #{e.message}"
-      Rails.logger.error msg
+    def skipped?
+      OpenProject::Configuration.ldap_groups_disable_sync_job?
     end
   end
 end
